Fix pytest-cov validation check.
